14_Documentation/djlibrary/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
djangorestframework==3.14.0
docutils==0.19
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.2
drf-yasg==1.21.5
